NXP-27209: Sonar cleanup

+++ addons/nuxeo-chemistry/nuxeo-opencmis-impl/src/main/java/org/nuxeo/ecm/core/opencmis/impl/util/SimpleImageInfo.java
@@ -153,7 +153,7 @@ public class SimpleImageInfo {
         return c;
     }
 
-    protected void skip(long n) throws IOException {
+    protected void skip(int n) throws IOException {
         long done = 0;
         while (done < n) {
             long num = in.skip(n - done);
